Refactor export module into SRP submodules
